fix costruttore SpellTroopSpawner
fix costruttore SpellTroopSpawner

+++ Model/SpellTroopSpawner.cpp
@@ -2,8 +2,8 @@
 
 /******************** CONSTRUCTORS/DESTRUCTORS  ********************/
 
-SpellTroopSpawner::SpellTroopSpawner(string n, unsigned int mana, rarity rar, unsigned int cLevel,string desc,unsigned int spellD,unsigned int crownTD,double rad,
-                                     unsigned int s,unsigned int th, double hxs, double dxs, unsigned int sdd, double r, unsigned int c,string timeSp):
+SpellTroopSpawner::SpellTroopSpawner(string n, unsigned int mana, rarity rar, unsigned int cLevel,string desc,double spellD,double crownTD,double rad,
+                                     double s,double th, double hxs, double dxs, double sdd, double r, unsigned int c,string timeSp):
 Card(n,mana,rar,cLevel,desc),
 Spell(n,mana,rar,cLevel,desc,spellD,crownTD,rad),
 Troop(n,mana,rar,cLevel,desc,s,th,hxs,dxs,sdd,r,c),timeSpawn(timeSp){}
